Renamed sbo-tester to operators-tester
